refactor: 컴포넌트 파일명 수정
