+++ src/components/Header/Header.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Header from '.';
-import CartTextButton from '../CartTextButton';
+import Header from './Header';
+import CartTextButton from '../CartTextButton/CartTextButton';
 
 const meta: Meta<typeof Header> = {
   title: 'Header',
