+++ src/components/ProductCard/ProductInfo/ProductInfo.tsx
