+++ src/components/ProductCardList/ProductCardList.tsx
@@ -2,7 +2,7 @@ import { useRecoilValue } from 'recoil';
 import { styled } from 'styled-components';
 import useLocalCart from '../../hooks/useLocalCart';
 import { fetchedProductListSelector } from '../../recoil/fetchSelectors';
-import ProductCard from '../ProductCard';
+import ProductCard from '../ProductCard/ProductCard';
 
 const ProductCardList = () => {
   const fetchedProductList = useRecoilValue(fetchedProductListSelector);
